fix(pg_list_tables): cast reltuples to float8 so estimated_rows is a number
pg_class.reltuples is float4. Casting it to ::bigint gives an int8 result
(OID 20), which node-pg parses as a string by default -- so estimated_rows
was returning "42" (a JSON string) instead of 42 (a JSON number), contrary
to the TypeScript annotation and every caller's expectation.

Cast to ::float8 instead: reltuples is already approximate, float8 just
widens the precision, and node-pg returns float8 as a JS number. Consistent
with the pattern used by other numeric fields in health.ts and explain.ts.

Adds a regression assertion to the integration test that verified
typeof estimated_rows === "number" against a live database.

+++ src/tools/schemas.ts
@@ -68,7 +68,7 @@ export const schemaTools = [
              WHEN 'p' THEN 'partitioned_table'
              ELSE c.relkind::text
            END AS type,
-           c.reltuples::bigint AS estimated_rows
+           c.reltuples::float8 AS estimated_rows
          FROM pg_catalog.pg_class c
          JOIN pg_catalog.pg_namespace n ON n.oid = c.relnamespace
          WHERE n.nspname = $1
